% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add_income}
\alias{add_expense}
\title{Add new items to a budget.}
\usage{
add_income(budget, name, amount, category, owner, months = 1:12)

add_expense(budget, name, amount, category, owner, months = 1:12)
}
\arguments{
\item{budget}{Budget object to add to.}

\item{name}{Name of the item.}

\item{amount}{Money amount.

  Can be either a single numeric, or a vector.
  If vector, its length should evenly divide that of `months`.}

\item{category}{Category of income or expense.

  For example `"housing"`, `"savings"`, etc.}

\item{owner}{Name of the person whose item this is.

  Used to generate individual budgets from joint budgets.}

\item{months}{Vector of integers corresponding to the months in which the
income/expense is active.}
}
\description{
Main functions for adding income channels and expenses to a budget.
}
\examples{
mybudget <- budget() \%>\%
  add_income("su", 5300, "grants", "Janus") \%>\%
  add_expense("rent", 3700, "housing", "Janus")

}
